pyproject.toml
[project]
name = "found-poll-buddy"
version = "1.0.0"
description = "Found. Poll Buddy is discord bot that allows you to create polls and vote on them."
authors = [
    { name = "The Found. Technology Authors", email = "foundtechnology@found-rg.com" },
    { name = "Abraham Murciano", email = "abrahammurciano@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "disnake==2.9",
    "asyncpg==0.29",
    "dateparser==1.2",
    "python-dotenv==1.0.1",
    "pytz==2024.1",
    "python-json-logger==2.0.7",
]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 101
preview = true
ignore = ["UP007"]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "**/stanza-cache",
    "**/models-cache",
    "scripts/",
]

[tool.ruff.lint]
exclude = ["*.ipynb"]

select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

exclude = ["*.ipynb"]